Improved plot_in_plane_functions

+++ docs/images/color_inversion_script.py
@@ -48,16 +48,23 @@ def invert_image_colors(image_path, output_path, new_color=(30,33,41)):  # Defau
             color_changed_image = change_black_to_color(inverted_image, new_color)
             color_changed_image.save(output_path)
 
-def process_folder(folder_path):
-    delete_previous_conversions(folder_path)
+def process_folder(folder_path, delete_old):
+    if delete_old:
+        delete_previous_conversions(folder_path)
     for filename in os.listdir(folder_path):
-        if filename.lower().endswith(('.png', '.gif')):
+        if filename.lower().endswith(('colorinverted.png', 'colorinverted.gif')):
+            print(f"Skipped {filename}, converted image already exists.")
+        elif filename.lower().endswith(('.png', '.gif')):
             base_name, extension = os.path.splitext(filename)
             inverted_filename = f"{base_name}-colorinverted{extension}"
             image_path = os.path.join(folder_path, filename)
             output_path = os.path.join(folder_path, inverted_filename)
-            invert_image_colors(image_path, output_path)
-            print(f"Processed {filename} -> {inverted_filename}")
+            # Only process if delete_old is True or the converted file does not exist
+            if delete_old or not os.path.exists(output_path):
+                invert_image_colors(image_path, output_path)
+                print(f"Processed {filename} -> {inverted_filename}")
+            else:
+                print(f"Skipped {filename}, converted image already exists.")
 
 # Example usage - replace 'path/to/folder' with the actual folder path
-process_folder('./docs/images/')
+process_folder('./docs/images/', delete_old = False)
